@@ -4,7 +4,7 @@ describe 'when a visitor goes to the homepage' do
   context 'the page they see if no pages have been created' do
     it 'has "Welcome to your new site" as the title' do
       visit '/'
-      page.should have_content "Welcome to your new site"
+      expect(page).to have_content "Welcome to your new site"
     end
   end
 
@@ -12,7 +12,7 @@ describe 'when a visitor goes to the homepage' do
     it 'is the first page of the first section' do
       test_page = FactoryGirl.create :lesson
       visit '/'
-      page.should have_content test_page.name
+      expect(page).to have_content test_page.name
     end
   end
 end
