@@ -13,7 +13,7 @@ describe Section, js: true do
     fill_in 'Name', with: 'Awesome section'
     fill_in 'Section number', with: '1'
     click_button 'Create Section'
-    page.should have_content 'Awesome section'
+    expect(page).to have_content 'Awesome section'
   end
 
   it 'displays errors if you try to save an invalid section' do
@@ -21,15 +21,15 @@ describe Section, js: true do
     visit table_of_contents_path
     click_link 'New section'
     click_button 'Create Section'
-    page.should have_content "Please correct these problems:"
+    expect(page).to have_content "Please correct these problems:"
   end
 
   it 'cannot be created by a student' do
     login_as(student, scope: :user)
     visit table_of_contents_path
-    page.should_not have_content 'New section'
+    expect(page).to_not have_content 'New section'
     visit new_section_path
-    page.should_not have_content 'New'
+    expect(page).to_not have_content 'New'
   end
 
   it 'can be edited by an author' do
@@ -43,21 +43,21 @@ describe Section, js: true do
   it 'cannot be edited by a student' do
     login_as(student, scope: :user)
     visit table_of_contents_path
-    page.should_not have_content 'edit'
+    expect(page).to_not have_content 'edit'
     visit edit_section_path section
-    page.should_not have_content 'Edit'
+    expect(page).to_not have_content 'Edit'
   end
 
   it 'can be deleted by an author' do
     login_as(author, scope: :user)
     visit table_of_contents_path
     click_link "delete_section_#{section.id}"
-    page.should_not have_content section.name
+    expect(page).to_not have_content section.name
   end
 
   it 'cannot be deleted by a student' do
     login_as(student, scope: :user)
     visit table_of_contents_path
-    page.should_not have_content 'delete'
+    expect(page).to_not have_content 'delete'
   end
 end
