@@ -3,7 +3,6 @@ require 'spec_helper'
 describe Chapter do
   it { should validate_presence_of :name }
   it { should validate_presence_of :number }
-  it { should validate_numericality_of(:number).only_integer }
 
   it 'validates uniqueness of name' do
     FactoryGirl.create :chapter
@@ -17,8 +16,8 @@ describe Chapter do
     last_chapter = FactoryGirl.create :chapter, :number => 9
     (2..8).each { |number| FactoryGirl.create :chapter, :number => number }
     first_chapter = FactoryGirl.create :chapter, :number => 1
-    Chapter.first.should eq first_chapter
-    Chapter.last.should eq last_chapter
+    expect(Chapter.first).to eq first_chapter
+    expect(Chapter.last).to eq last_chapter
   end
 
 end
